Add Macwire components
This does not have the perfect organization and could be improved. But
it works and shows how to translate pac4j security module to Macwire.

Most of the translation here involved moving from Guice `@Provides`
annotated methods to components wired (sometimes manually) by
Macwire.

We are not yet activating MacwireApplicationLoader yet, but that is the
easiest part.